Suppressed Xcode compiler warnings.
Added initialization of some variables. This is not a real uninitialized
variable problem - when reading/writing valid charstrings, these
variables would all have been initialized before being used, but the
compiler can't tell that.

+++ libpsautohint/src/fontinfo.c
@@ -34,7 +34,7 @@ GetKeyValue(const ACFontInfo* fontinfo, char* keyword, bool optional,
     fontinfostr = GetFontInfo(fontinfo, keyword, optional);
 
     if ((fontinfostr != NULL) && (fontinfostr[0] != 0)) {
-        *value = atol(fontinfostr);
+        *value = (int32_t)atol(fontinfostr);
     }
     return;
 }
